requirements.txt
certifi==2018.4.16
chardet==3.0.4
cryptography==1.7.1
enum34==1.1.6
graphviz==0.8.3
idna==2.6
ipaddress==1.0.17
keyring==10.1
keyrings.alt==1.3
-e git+https://github.com/apache/incubator-mxnet.git@25720d0e3c29232a37e2650f3ba3a2454f9367bb#egg=mxnet&subdirectory=python
numpy==1.8.0
picamera==1.13
pyasn1==0.1.9
pycrypto==2.6.1
pygobject==3.22.0
pyxdg==0.25
requests==2.18.4
RPi.GPIO==0.6.3
SecretStorage==2.3.1
six==1.10.0
urllib3==1.22
